+++ app/constants.js
@@ -3,4 +3,6 @@ exports.ROLES = {
   default: 'regular'
 };
 
+exports.TOKEN = 'token';
+
 exports.ALBUM_SUPPLIER = 'https://jsonplaceholder.typicode.com/albums';
