+++ app/routes.js
@@ -5,7 +5,9 @@ const { signUp, signIn, listUsers, signUpAdmin, getAlbums } = require('./control
 const {
   validateSignUpDTO,
   validateSignInDTO,
-  validateGetUsersDTO
+  validateGetUsersDTO,
+  validateIdentity,
+  isAdmin
 } = require('./middlewares/validators/user');
 const { validateToken } = require('./middlewares/validators/token');
 const { jsonSwagger } = require('./controllers/documentation');
@@ -15,8 +17,8 @@ exports.init = app => {
   app.get('/docs', redoc({ title: 'API docs', specUrl: '/docs/swagger' }));
   app.get('/health', healthCheck);
   app.get('/users', [validateGetUsersDTO], listUsers);
-  app.get('/albums', [validateToken], getAlbums);
+  app.get('/albums', [validateToken, validateIdentity], getAlbums);
   app.post('/users', [validateSignUpDTO], signUp);
   app.post('/users/sessions', [validateSignInDTO], signIn);
-  app.post('/admin/users', [validateToken, validateSignUpDTO], signUpAdmin);
+  app.post('/admin/users', [validateToken, validateSignUpDTO, validateIdentity, isAdmin], signUpAdmin);
 };
